[kernel][vm][ppb] StackOwnedLoanedPagesInterval
This RAII class will allow a thread to wait on a loaned
page to no longer be owned by another thread's stack.

The OwnedWaitQueue is used for proper priority inheritance
so that the waiting thraed doesn't get stuck waiting for
a lower-priority thread.

This CL stops short of making the PageQueues changes needed
to keep a backlink set while a loaned page is in a VmCowPages.
That'll be in a later CL.  For now, there won't be any loaned
pages so the clearing of the backlink sometimes while a page is
still in the same VmCowPages is still ok for now.

Tested: vm-stress-test, CQ, vmo_stack_owned_loaned_pages_interval_test
Change-Id: I20a86e848877fe2447979e632fb72f9498cfc52f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591623
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>