[usb][xhci] correct segment_index_ logic
segment_index_ is written to the DESI bit of ERDP and should be the last
3 bits of the offset of the current ERST entry (see 5.5.2.3.3 in XHCI
spec). It is used to quickly determine if the Event Ring is full.
Originally, we were increasing segment_index_ only when we wrapped
around from the last buffer to the first buffer, but we should be
incrementing segment_index_ every time we move to the next buffer (i.e.
segment).

Test: tested on NUC (17 hours and mouse is still working)
      tested on Atlas (mouse isn't laggy anymore)

Bug:81262
Bug:78895

Change-Id: I1027184e8889bd9d5e2b6885e270fa108caed3cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599861
Reviewed-by: Sam Hansen <hansens@google.com>
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>