[packet] Relax BufferView lifetime for &mut &[u8]
In our implementation of `BufferView` for `&'b mut &'a [u8]`, we
previously used the more restricted lifetime `'b`. Concretely, we had
the impl:

  impl<'b, 'a: 'b> BufferView<&'b [u8]> for &'b mut &'a [u8]

This is problematic because it means that, if a caller holds a reference
with lifetime `'b` to an object which owns a buffer with the longer
lifetime `'a`, the caller can only use `BufferView` with the shorter
lifetime `'b`. In practice, this has consequences for situations like
parsing a packet from a buffer which is stored in a local variable. If
the buffer's storage has a longer lifetime, we want to be able to parse
and produce a packet object with that same lifetime, but this
`BufferView` impl means that we can only produce a packet object with
the shorter lifetime of the local variable.

We change this impl to use the longer lifetime `'a`:

  impl<'b, 'a: 'b> BufferView<&'a [u8]> for &'b mut &'a [u8]

This solves the problem, and allows us to parse a packet object with a
lifetime which outlives the lifetime of the local variable used to store
the buffer.

Change-Id: I278783a9b573308d0fbeb74135622e97d6121ef6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594026
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>