[build] Add vpython3 to tool_paths.json
So the sdk recipe can use vpython for invoking SDK scripts. I was hoping
to be able to stop using vpython entirely, but that causes some
breakages on Mac, so instead I'm going to just make sure that the recipe
uses the versions of vpython3 and python3 that are included in the
checkout.

Bug: 77963
Change-Id: Ieac569806f6b0627ccc16297e19945d16b87ea51
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601202
Commit-Queue: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>