[packet-formats] Add TCP/UDP into_body methods
Add methods to `TcpSegment` and `UdpPacket` which consume the object by
value and return the underlying body. The returned body has the same
lifetime as the buffer that the packet object was parsed from. By
contrast, the existing `body` methods (which take `&self`) return slices
with the (possibly more restricted) lifetime of the packet object
itself.

Change-Id: I83c422887ba9cc6ba36e7f2f3e9fe6743b3c4ef6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594027
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>
Reviewed-by: Nick Brown <nickbrow@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>