[timekeeper_lib] Replace zx::time_utc with a library-defined type.
Zircon previously exposed a UTC clock but this was removed with the
zx_clock_get syscall June 2021. We now need to clean up to remaining
references to UTC in zircon including the ZX_CLOCK_UTC constant and the
zx::time_utc type.

The zx::basic_time<> type will still exist and so can easily be used to
define a replacement type. This CL defines a replacement in
lib/timekeeper and uses this new type for in-tree clients. The new type
is API compatible, allowing a soft transition for out-of-tree usage of
lib/timekeeper.

Bug: 86976
Change-Id: Iea1f3723bfc36c7e0decb232dbd4509730aa0603
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599944
Reviewed-by: Andy Weiss <dragonbear@google.com>
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>