[ffx][emu]Simplifying FfxConfigWrapper and test usage
The previous implementation used an async trait, and 2 implementations
of the trait, one for tests which was backed by a hashmap, and one for
non-test usage which called ffx_config methods directly.

This change merges the two and removes the async trait. It is much more
simple to understand what is going on, and removes the complexity of
async_traits and sharing them across threads (which I encountered in
a prototype).

Change-Id: Ieeb75161fe5f8870135ed3aa975ba9c19bb2feaa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/600226
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>