[bt][a2dp] Don't start player until data arrives
In some cases when a source stream is connected, it is suspended without
sending any data.  This causes us to start up a Player and immediately
tear it down, which is undesirable.

Instead start the player once the first data arrives in the data stream.

Add an estimate of the total streaming seconds to the inspect data for
a DataStreamInspect.

Test: tests updated
Test: fx test bt-a2dp-tests bt-a2dp-encoding-tests
Test: fx test bluetooth-crate-tests bt-rfcomm-tests
Bug: b/203686738

Change-Id: Ib4890d3a4f70d9a210239fc22e62ef405070e3cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599361
Fuchsia-Auto-Submit: Marie Janssen 💖 <jamuraa@google.com>
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>