[i2c-hid] Migrate i2c-hid.cc to new WireSyncClient API
This is part of a migration to stabilize the LLCPP API. In this step, we
are aligning the interface of WireSyncClient and WireClient. The main
changes to WireSyncClient are:

- Similar to WireClient, use the "->" operator to make FIDL calls with
managed memory allocation.

- Similar to an earlier change in `fidl::WireCall(...)`, use the
`client.buffer(arena_or_buffer_span)` accessor to make FIDL calls with
caller-controlled allocation.

- Removing accessors that expose a raw pointer/reference to the
channel, instead providing a `Bind` function to initialize a default
constructed client, and a `TakeClientEnd` function to de-initialize it.

- When binding a client endpoint, the client endpoint must be valid.

See https://groups.google.com/a/fuchsia.dev/g/fidl-dev/c/nF72xpPL4gU

Bug: 87683
Change-Id: I8638c4dc051cbabf7deb8deffefd810b1c171d15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/600584
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>