[component-events] rename `moniker` to `moniker_regex`
The `moniker` function in `component_events::EventMatcher` takes a
regular expression to match against moniker in received events, as
opposed to a string that's expected to match the relative moniker
exactly.

Rename this function to `moniker_regex`, to better convey to the user
what the function does. Also rename `monikers` to `monikers_regex`, for
the same reason.

Change-Id: I64f12ef30093e3c2af12635a1f179ead5519a745
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601261
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>