[bt][rfcomm-integration-tests] Fix flake in test
The RFCOMM integration tests occasionally flake when validating that
data can be written to one end of a RFCOMM channel and read from the
other end.

Fix this flake by reading in the receiving end until data is produced
(or a test-defined timeout is hit). This matches the behavior of the
bt-hfp-audio-gateway-integration-tests which experienced similar issues.

Bug: 82484
Test: fx test bt-rfcomm-integration-tests
Multiply: bt-rfcomm-integration-tests
Change-Id: If34f779d38d533508e1e843d6a980d4d8341ed1b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599271
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>