[assembly] Add signed support for images manifest
The zbi in the images manifest may be signed, and we need to pass this
bit of info to the create-update operation, because the entry in the
UpdatePackage depends on it (zbi vs. zbi.signed).

Also add the signed zbi to the images manifest, instead of the unsigned
one.

Bug: 84904
Change-Id: I40caf892cb2291623aacd22b32fa0902b1693308
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601011
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>