[build] Ensure components and packages are only defined in the default toolchain
There are many occurrences where fuchsia_component() and other
package or component related templates are invoked in the non-default
toolchains.

For example, a compiled_action() points to a tool that is defined in
a BUILD.gn file that also declares a fuchsia_unittest_package().

The latter will be expanded in the context of the host toolchain (even
if the result is never added to the final build graph), creating
un-necessary targets and dependency chains, and GN work.

Another example is a library that needs to connect to a component
at runtime, and lists it in deps or data_deps.

This CL ensures that all these declarations end up redirecting to
the appropriate default_toolchain. What the change amounts to is
something like:

  fuchsia_unittest_package() {
    if (current_toolchain == default_toolchain) {
      ... # old template content
    } else {
      group(target_name) {
        testonly = true
        deps = [ ":$target_name($default_toolchain)" ]
      }
    }

Bug: 87739

Change-Id: I26b4f36919342f6ceb37db9808be7069d41aee6c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598463
Fuchsia-Auto-Submit: David Turner <digit@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: David Fang <fangism@google.com>