[test_manager] Wait for client to close suite controller
Holds the SuiteController channel open instead of closing it after
all events are drained. This is a workaround for an issue where overnet
returns ZX_PEER_CLOSED if the server returns an event, then immediately
terminates the channel. See fxbug.dev/87890.

Note that the SuiteController API explicitly states that the server
will drop the channel after events are drained. This breaks that API,
however, none of the clients currently rely on this behavior. We'll
need to remove this workaround once the bug in overnet is corrected.

Multiply: test_manager_test

Bug: 87976, 87890

Change-Id: Ib357e133ac10a60eab2c4e512a8f56d4b225d5a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601524
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>