[zircon][vm] Get backlink until after pages removed.
To chase down a page that needs to be reclaimed, we need
to be able to get a RefPtr<> tally on the owning
VmCowPages until after the page has been removed from the
owning VmCowPages.

This avoids the page reclamation loop spinning between
VmCowPages refcount 1 to 0 and ~VmCowPages removing the
pages.

To do this, we separate the VmCowPages and its container
to allow each to have its own refcount.  The VmCowPages
refcount can drop to zero while the container's refcount
is still 1.  During this time, we can still get a ref to
the container, and via that ref we'll be able to replace
the page.  Before VmCowPages removes its 1 ref tally on
its container, it has fully removed all pages to FREE
state.

Change-Id: I7c51bfa6392f1f6fd26ea2766ae3cb8f15ca2d26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601301
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>