[ffx] Reorganize the emu plugin to support engines
We've had a lot of conversations about how the emultator plugin should
be organized lately, as we prepare to add multiple instances and
multiple backing engines. We settled on an object-oriented design where
each engine is its own type that implements a common interface, and the
engine used for a given instance comes from the manifest type selected
during launch.

This CL introduces a new layout for the code in the emulator plugin
source tree which reflects the engine-based design. The existing code
becomes the "vdl" engine, which uses device_launcher as the backend.
Common code, such as the port_picker library, remain in the common
library.

A few skeleton types are also added, which will be populated and
plumbed through in the next CL, for the common config types that will be
sent to each engine.

Change-Id: I25b024e89b245673864fef0e83efb204c80868b6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602565
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>