[flatland] Defer response to ChildViewWatcher.GetViewRef().
The previous behavior was to respond to GetViewRef() as soon as the
ObjectLinker connected the view/viewport.  This resulted in the
possibility of the parent immediately attempting to set focus on the
child view, before the child appeared in the ViewTree, and therefore
focus would not be set.

Fixed: 81850
Change-Id: I9c9658d170b00181f0b2363eaf2bcf8a378d776e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601921
Commit-Queue: Josh Gargus <jjosh@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
API-Review: Jaeheon Yi <jaeheon@google.com>