[starnix][msync] Add validation to unimplemented msync
The gvisor mremap tests use msync as a way to check whether a page is
mapped. They expect an ENOMEM error when msync is called on an address
range that has one or more unmapped pages.

This change implements this validation, but still logs that the syscall
is unimplemented because it doesn't actually do anything besides the
validation.

Change-Id: I15012f47047193738e8564e884c6b40ef3265bc0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602811
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>