[realm_builder][rust] reenable altered_echo_client_arg
The `altered_echo_client_arg` integration test in realm builder was
modified in fxrev.dev/599742, and due to an oversight failed to give the
LogSink protocol to `echo_client`, which resulted in the test failing.

Route the LogSink protocol to `echo_client`, and reenable the test.

MULTIPLY: fuchsia-component-test-tests
Fixed: 87870
Change-Id: If293b69d80121980615e4c6575ab2b440118dc72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602002
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>