[fidlfmt] Adjust comment indentation algorithm
The algorithm intends that a standalone block comment must be
indented to match the indentation of either it's immediately preceding
or immediately following line, whichever is greater. The previous iteration
of this code, however, was only indenting the comment to match the
indentation of the following line, completely ignoring the preceding
one.

Additionally, `find . -iname '*.fidl' -not -path "./out/*" -not -path
"./prebuilt/*" | xargs -I {} fx format-code --files={}` was run to
ensure that all in-tree files incorrectly formatted due to this bug are
now correct.

Test: fidldev test fidlc
Change-Id: I6d318ba15a502322290c056d09f42b1a5adfe6f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602262
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Pascal-Louis Perez <pascallouis@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>