[json_validator] Create valico-based validator.
The existing json_validator tool suffers from a fundamental bug in
which it ignores schema references. The Rust library valico is
already used in-tree, and it provides a relatively easy option
for a replacement, as long as clients use Draft 7 of the schema
specification.

Bug: 87807
Change-Id: I8100b034d7b86d385102ed7d303b8ac88e301080
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601530
Commit-Queue: Onath Claridge <claridge@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Shai Barack <shayba@google.com>