[fidl] Fix typos and markup for block device FIDL.
Declare `BLOCK_VMOID_INVALID` to be of type `Vmoid` (= `uint16`), not
`uint32`.

Set type of fields in `block_stats_t` from `size_t` to `uint64_t` to
match the FIDL. `size_t` is technically not guaranteed to be `uint64_t`,
although as a practical matter it is on all Fuchsia systems.

Update the comment explaining the `flags` field of `alloc_req_t`. It
took some spelunking to find the actual live definition. It is still not
super clear; perhaps the flag name(s) should be in an `enum` with a
suitable name, or prefixed with `ALLOC_REQ_T_FLAG_` or similar.

Bug: None
Change-Id: I3b8fdcab5bcdcda813f35c0347dae70548db4ba9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602109
Commit-Queue: Chris Palmer <palmer@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>