[ffx] Remove package serving from ffx emu
Previously, ffx emu was built to start the package server for the user
alongside the emulator. As we redesign the launcher, we want users to be
responsible for starting it themselves. In August, we made running
without the package server the default and left a note to completely
remove it after a month. It's now been 2.5 months, so the time has come.

This CL removes the package serving flags from ffx emu, making it
impossible for a user to start the server from here. They must use
something like `fx serve` or the upcoming `ffx repository` instead. This
will not impact infra tools that are already starting it themselves or
using other launchers.

Bug: 82804
Change-Id: Iba4532eeb14b3516c1095a917b629dfab572a6b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601541
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>