[audio][tools] signal_generator PTS units
The signal_generator CLI utility previously was capable of setting
timestamps on Play (ref and/or media independently) and/or on packets,
as well as setting a PTS continuity threshold, It was NOT previously
capable of manipulating the PTS unit definition (i.e., it always used
the frame-number as this unit).

With this CL, signal_generator can specify an arbitrary definition
for the PTS unit. It also correctly reflects the default PTS unit of
1 nsec (as AudioRenderer does) if no PTS unit is specified.

Additionally, we rename timestamp-related cmdline flags for clarity.
- When calling Play(), '--ref-start' (not '--ref') and '--media-start'
  (not '--media') indicate that specific values should be set.
- To apply PTS values to packets, use '--packet-pts' (not '--pts').
- Set the PTS continuity threshold duration with '--pts-threshold'
  (not '--threshold').
- Use new flag '--pts-unit' to specify the PTS unit. If absent, the
  default unit of 1'000'000'000/1 (1 nanosec) is used in the Play()
  call and/or on packets, as needed.

This functionality was added to signal_generator to better debug
issues involving nanosecond-timestamps, such as the bug below.

Bug: b/204603590

Change-Id: Ia0ec9b188d70a133139800f25ff9e62f47d09fba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/600964
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>