[remote-control] Close tunnel if one side closes
While previously we would shut down the tunnel if the zx socket was closed,
we wouldn't shut down the zx-to-tcp task if the tcp socket was closed. This
could waste resources, as it'd leave the zx socket up and running, and thus
the tunnel to the host.

This fixes that by using emitting a () from the tcp-to-zx task over a
channel to the zx-to-tcp task, and adds some tests to verify this logic.

Change-Id: I6e53114add6e77b7fe8dd919db8bdcde56c18382
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/603274
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>