[cflags] Make FIT_DECLARE_THREAD_CHECKER -Wextra-semi compliant
This macro emits either a member declaration or nothing
depending on NDEBUG.

The declaration needs to terminate with a semicolon,
whereas an empty definition cannot have a dangling
semicolon after it or it triggers -Wextra-semi.

This leaves two options:
1. Change all usages of FIT_DECLARE_THREAD_CHECKER(...)
to not terminate in a semicolon.
2. Surround the no-op definition of FIT_DECLARE_THREAD_CHECKER
with a suppression for -Wextra-semi.

I chose the second option, for consistency with
FIT_DCHECK_IS_THREAD_VALID which is both terminable
and terminated with semicolons at every usage site.

Unfortunately the same trick used by
FIT_DCHECK_IS_THREAD_VALID isn't valid here,
because ((void)0) only works in a function body
and not in the context of a declaration.
There is no no-op declaration spelling that can be
terminated with a semicolon.

Bug: 56257
Change-Id: Ie94ac86209264fce16d96cfc0063a7f755d9739d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/603581
API-Review: Shai Barack <shayba@google.com>
Fuchsia-Auto-Submit: Shai Barack <shayba@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Commit-Queue: Shai Barack <shayba@google.com>