[core] Integrate core realm proxy
This change deploys the core realm proxy that was implemented in
fxr/596663. This changes the topology so that `core` itself is
responsible for bridging protocols between v1 and v2, rather than
`appmgr`. This provides a smoother developer experience for migrating
`sys` components, by creating the following changes when routing a
protocol between v1 and v2:

- An `offer to #appmgr` declaration in `core.cml` or a core realm shard
  becomes `use from #appmgr`.
- An `offer from #appmgr` declaration in `core.cml` or a core realm
  shard becomes `offer from self`.
- Changes to `appmgr.cml` are no longer needed. (Also, this means paths
  no longer need to be prefixed with `/svc_from_sys`, which cuts down on
  boilerplate.)
- Changes to `verify_routes_exceptions_allowlist.json5` are no longer
  needed. (A followup change will remove the exceptions that already
  exist.)

Bug: 86886

Change-Id: I7fc6fda98ce480e5d20eb075553a0e70c2b79fb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595041
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>