[iwlwifi] Better handling the time event removal
The original code removed the time event during the associaion stage.
This is okay for most cases, but not always correct.

The correct way to handle this is relying on the
TIME_EVENT_NOTIFICATION from the firmware, which tells us when the
time event starts and ends. So, the new code re-enables the driver
code to correctly handle the notification.

Bug: 87392
Tested: Added a test case to cover it.
        On 9260, with v38 and v43 firmware. Both works pretty well.

Change-Id: Ia27df69066e5b463a4c01ed270f1722d62d9a14e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601531
Commit-Queue: Louis Yung-chieh Lo <yjlou@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>