[netstack] Add bridging integration test
The test creates N+2 netstacks. The 2 netstack that are always present
are named `gateway` (not actually a gateway since it doesn't route
packets to the outside) and `switch` (not actually a switch, but just
a Netstack on which interfaces are bridged together).  `switch` and
`gateway` are connected via a network. In the simplest testcase,
a single other Netstack is created (named `host1`), connected to
`switch` via a network. The interfaces on `switch` are then bridged,
and the test verifies that `host` can acquire addresses and communicate
with `switch` and `gateway`.

There are also test cases to ensure that link/admin status changes on
the bridged ports do not affect communication on the bridge.

Also add a test which ensures that removing a bridge does not
unintentionally enable an attached interface if it is disabled.

Multiply: netstack-bridge-integration-test
Test: fx test netstack-bridge-integration-test
Fixed: 85075
Change-Id: Ib52f64f3195120c7f276be9f234e381c28bb2de0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585482
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Tony Gong <gongt@google.com>