[zircon][vm][ppb] PhysicalPageBorrowingConfig
This class is for holding enable/disable state for physical
page borrowing.

As of this CL, borrowing is still disabled and not all the
needed code to borrow is merged.  Later we can enable
borrowing via the config defaults.  See also the TODO
comment in physical_page_borrowing_config.h.

Change-Id: If509b7b4cc7177771cc47c218a4f9e68f51d0f7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601745
Commit-Queue: Dustin Green <dustingreen@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>