[llcpp] Use VTable rather than transport type in message classes
VTable is now used to indicate the type of the transport and allow
easier lookup of fields stored on the VTable. The motivation for doing
this is to avoid a dependency in llcpp_transport.cc on the all of the
transports.

Bug: 85734
Change-Id: I5df3612a97e1e65e6b357f24d1bdd3da4e4f5322
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602568
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>
API-Review: Pascal-Louis Perez <pascallouis@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>