[opencl] Create OpenCL loader FIDL
This uses the same approach as the vulkan loader.

The interface is discoverable. The OpenCL drivers on Fuchsia implement this
interface is they support being loaded via the OpenCL-ICD-Loader.

The OpenCL-ICD-Loader on Fuchsia searches for this interface to find OpenCL drivers.

fixes: 88249
Change-Id: I6e6eb9ced1c6d20a22a16c43e81d57ef2ac00ef5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/603882
Commit-Queue: Jamie Howarth <jhowarth@google.com>
Reviewed-by: John Bauman <jbauman@google.com>
API-Review: John Bauman <jbauman@google.com>