[llcpp] Use transport vtable implementation for message write, call and read
This CL:
- Calls vtable->write() rather than zx_channel_write_etc for message
write, similar for call and read
- Adds CallOptions, ReadOptions, WriteOptions to respective message and
transport functions. This was needed for read() now that it is
transport agnostic but the same change was applied to Write and Call
so they parallel one another.
- Removes EncodeFlags / DecodeFlags. These might be needed at some point
but there are no short term use cases.

Bug: 85734
Change-Id: Id3b128c3d1e74b887bed7e8103683c20cd2cd5ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602569
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
API-Review: Ian McKellar <ianloic@google.com>