[kernel][vm] Commit and MapRange do not dirty
When a pager-backed VMO is committed, its pages are populated, but not
dirtied. When a pager-backed VMO is read-write mapped with MAP_RANGE,
the mapping is populated, but the pages are not dirtied. Both these
cases should not generate ZX_PAGER_VMO_DIRTY requests.

Also fix up mappings for pager-backed VMOs so that they are mapped
initially with the write permission removed, so that future writes can
be trapped and the dirty state updated.

Test: core-pager

Bug: 63989
Change-Id: I9a56695f9943b481246ab934ff2c8397bff1c889
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602784
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>