[test_diagnostics] Don't panic on channel closed
Due to recent change to pipeline requests, the channel may be closed.
This occurs because the client closes the channel and drops
outstanding requests when it receives the first error. When the server
attempts to respond with a second error, the channel is closed.

Change-Id: I4de5b874ea357561a953f1d80fbc91f5e0851145
Fixed: 88295
Multiply: test_diagnostics_host_unittests
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604223
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>