[setui] Swap all stash commits to flushes
This is a temporary and easily cherry-pickable solution to switch all
commits in SetUI to run flush instead. If the device runs out of space
when committing, stash will close the connection, which causes that
particular setting to be stuck until the device reboots.

Using flush will allow SetUI to continue to function and keep the
settings values in-memory, and return to normal functionality once the
device is no longer out of space.

Bug: 88349
Test: fx test setui_service_tests
      manually tested on device by continually cycling flashlight and
      volume slider
Change-Id: Idc81e22a78f9a89e664a33338b484dd4cfb76174
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604647
Commit-Queue: William Xiao <wxyz@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>