[brcmfmac] Replace scan_request with boolean ongoing_scan field
The pointer copied into the scan_request field of struct
brcmf_cfg80211_info is not owned by brcmfmac, so a proper copy of the
data it contains should be performed if we wish to retain it. Since
this field is only used as a boolean, instead of fixing the ownership
problem, the field is replaced by an actual boolean instead.

Bug:
Test: Pure refactor
Change-Id: I0c5ba63d780fc23546c3c06c40aa6b6a0a316d1d