[fxfs] Fix some flush issues
Keeping track of the minimum size was proving complicated and it had
bugs. This CL simplifies things by removing the minimum-size tracking
and making sure that we commit shrinking a file before we grow the file
again.

This CL also addresses some issues with reservation tracking by the
cache. Keeping the reservation sizes accurate whilst allowing writes and
flushes at the same time is quite difficult: if a write arrives after a
flush has started, then the amount we want to reserve might depend on
whether the write is in a range that is still to come or in a range that
has already been flushed. To simplify things, this changes flushing so
that it always flushes *all* the outstanding data and then, to make sure
transactions cannot get too big, we make sure to limit the amount of
outstanding data for a particular file.

Fixed: 87196
Change-Id: I711a0a98a4aeec9fe8b279853e26121b7e9db35f
Multiply: fxfs-truncate-large-tests 100
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601621
Commit-Queue: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>