[fxfs] Make opening stores thread-safe
A previous CL erroneously removed thread-safety for opening stores which
was believed to not be necessary, but it turns out that it is. The
graveyard needs access to stores for its initial reap process and at the
same time stores can be opened for mount.

Bug: 88401
Multiply: fxfs-corrupt-tests 100
Change-Id: I7d6cd37e60423629dd10313ea94172e538cc5a2d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605101
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>