[fidlc][c] Emit coding tables for request/response payload types
This CL removes the conditionals in fidlc suppressing coding tables for
request/response payload types.

Given a method `mylib/Foo.Bar`, the C request type for the method is
`mylib_FooBarRequest`. However, this conflicts with the name of the
request/response payload type that is synthesized by fidlc. This CL
renames the message types (header + payload) to RequestMessage, and
similar for responses and events.

Some low-level libraries that are using the C bindings in a bare-metal
way are updated. The coding table is an internal API, hence should be
API compatible to rename.

Change-Id: Ibeb9f650cbd008e1d4770a4713e43f063a2356f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604069
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>