[ffx] Add additional fields to virtual_device.json
The ffx emu tool is building out support for the product bundle
manifest, including a new virtual device specification. This drives
the configuration of the emulator.

This CL adds the fields that previously were defined in the DeviceSpec
struct to the manifest and its schema, and builds on the PBM
functionality to parse the manifest and pass it to the emulator engine.
It also introduces the trait for emulation engines, and embeds the
existing device_launcher-based code into a VDL engine.

Change-Id: If22eb3f0bf51b4f2b17f3fd16fc2111e8ddc79ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604283
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Misha Gridnev <gridman@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>