[blobfs] Reduce compression factor for multithreaded test.
The test sometimes takes too long to run for the 4 thread case. Since
the amount of data generated is scaled up based on thread count we end
up compressing 20M of data in serial, and that compression goes really
slowly in a lot of builds. We don't really care about testing the
efficiency of compression here, just thread interactions. So lowering
the compression factor to speed up the data generation at the beginning.

Bug: 87594
Change-Id: I1a09dc867b17b1c646bd50991518f9b0c98e4387
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604961
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Martin Lindsay <mlindsay@google.com>