[ffx] Add a README to document config_types design
In our redesign of the ffx emu tool, we are introducing a strict
separation of roles in the data types used by the emulation engine
interface. Those roles are expected to be clear to a regular user, but a
developer who's less familiar with the system may be confused by our
intentions.

This CL adds a README document in the config_types crate which explains
the design philosophy around the types and their roles. We expect this
will reduce churn during code reviews in the future, by providing
contributors with a solid foundation to understand the design.

Change-Id: I6360616851ce4ec371b5480db3c2c2330b5c44ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605303
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>