[ffx] Rename the common lib build label
In our recent refactor, we renamed all of the build targets in the ffx
emu plugin so that the build label matches the crate label. We left out
the common lib, because we had another plugin which depended on it. Now
that that dependency is removed, we can finish the refactor safely.

This CL renames the build target fro "lib" to "ffx_emulator_common", to
match the crate label. It also updates all of the references to that
target accordingly.

Change-Id: I7f342e1e3f7bc3dd70608df9ae0df31658e7427f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605304
Fuchsia-Auto-Submit: Daniel Dougherty <doughertyda@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>