[llcpp] Wire through transport-specific context
The driver transport needs to pass an arena from user-facing
method calls down to transport write calls and in the other
direction from transport read calls to user-facing method
responses.

This CL introduces mechanism to pass through a transport-specific
pointer through these layers.

No test is included in this CL, but it will be covered in an
upcoming integration test. The reason for not adding the integration
test now is that more work needs to be done elsewhere to get
to the point where an integration test is feasible.

Change-Id: I35d356de1e638e395eb2dab746b4ba8eb2dc194f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604742
Reviewed-by: Pascal-Louis Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Pascal-Louis Perez <pascallouis@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>