[virtualization] Update termina build script
This build updates the sonames for EGL and GLES to be
compatible with ldconfig in the container.

Also minigbm is built from source.  We're applying patches
for now but expect to land these after consultation with minigbm
owners.

A magma source area serrves to build libmagma there, making it
available in the sysroot so mesa and minigbm can both link it.

Bug:82886

Change-Id: Ida7b0191f9cba130a333b65a36507f25a7a399b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605823
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: Craig Stout <cstout@google.com>