[ffx] Strip the legacy code out of ffx emu
We've been working around the legacy VDL code in ffx emu for a while
now, and it's starting to cause more work without providing any
usefulness. Existing users can use fvdl or fx emu, and since we're still
experimental, we have the ability to make sweeping changes.

This CL strips out the VDL code, leaving only the existing StartCommand,
and some of the new framing we've been adding lately. We're going to
build out our tools without trying to fit the old code in the new
framework. We can refer to the fvdl code base if we need to reference
any old code.

Change-Id: Ic0265f4231d2ed8d7bf096f33e80df43e9ce9fa5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606021
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>