[base-resolver] Disallow package hashes in URL
Currently package hashes are ignored, which can lead to unexpected
package resolution.

This change causes base resolver to return an error when resolving a
component URL with a package hash.

In the future, it might be useful to resolve URLs with package hashes if
they match the package's hash in base, but this isn't needed right now.

Bug: 78072
Change-Id: I23548bf23cc98c21fc0fcc228dc26c0f67c0100e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605503
Reviewed-by: Dan Johnson <computerdruid@google.com>
Commit-Queue: Adam Lesinski <adamlesinski@google.com>