[archivist][tests] rename "add_eager_component" to "add_eager_child"
The realm builder library is now using "child" instead of "component".
The "add_component" function was renamed to "add_child" in a prior
commit, but "add_eager_component" was missed in this migration.

Rename it to align with the realm builder library it's wrapping.

Change-Id: I9b37f1bce53ad15d5a0eac5c6418473025b4f752
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605902
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>