[ffx][wlan] Adds commands to control SMEs.
This change adds direct control of WLAN client, AP, and
mesh interfaces.  Two client commands have been
deliberately omitted in this first translation pass.  The
client scan and client join commands are not easily
portable at the moment as both require the parsing of scan
results.  The libraries that enable this parsing rely on
banjo bindings in the SDK which appear to be absent from
the host build.

Bug: 87486
Test: fx test ffx_wlan_dev_args_test
Test: ffx wlan dev client disconnect
Test: ffx wlan dev client wmm_status
Test: ffx wlan dev ap start -s asdf -c 11 -i 1
Test: ffx wlan dev ap stop -i 1
Test: ffx wlan dev mesh join -i 1 -m asdf -c 1
Test: ffx wlan dev mesh leave -i 1
Test: ffx wlan dev mesh paths -i 1
Change-Id: Ib7580351645765d8284728764da71894fbcaa4f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606202
Commit-Queue: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>