[gidl][c] GIDL C encode success tests
This C adds support for encode success GIDL tests. It uses LLCPP to
linearize input objects for use in the C encoder.

Encode failure tests are not part of this CL because for failure tests,
the same linearization process typically produces failures in LLCPP
making it not possible to test the C encode failures.

Change-Id: I4d61cc0214f74bb732b45fa1746519d577a22b7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606461
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>