[minfs] Remove unused build arg.
The minfs_board_maximum_data_bytes was used to tell appmgr the maximum
size of the partition which it would use in cache clearing logic. This
was a temporary workaround for minfs not accounting for the partition
limit in its size calculations.

The minfs size calculations have been corrected and this value is no
longer used.

Bug: 85718
Change-Id: I5c6510bf9986a65c52649059fc5f4c26a25b3d7d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/606086
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>