[rust-3p-crates] num's `use_addcarry` only works on x86/x64
It appears that when num got updated, we enabled the `use_addcarry`
feature, but didn't end up re-generating the BUILD.gn. When re-run, this
will fail to compile on aarch64, because the `use_addcarry` feature only
works on x86 or x86_64.

This patch changes the configuration to only set this feature if we are
running on the appropriate target.

Change-Id: I688ff532bc1eebd5ddd86091e984167142f28f52
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/607188
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>