[virtualization] Implement fuchsia.net.virtualization in FakeNetstack
Implement the new fuchsia.net.virtualization/Control FIDL protocol in
FakeNetstack. This provides a fake implementation of the protocol that
allows a real virtio-net device to connect, and then allow tests to
send/receive packets to guests.

To support a soft transition to the new interface, the implementation
currently supports both the new (fuchsia.net.virtualization.Control) and
old (fuchsia.netstack.NetStack) FIDL interfaces. When a test attempts to
read/write to a network device, the read/write will succeed if the
network device was attached via either the new or old protocol.

Bug: 87034

Change-Id: Ia4779e513a54e120fb5ec4a1d59e74c03254cb71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601681
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>